A lot of smaller changes. Added replay function to the high scores, so now you can see the best scores live!

+++ src/logic/subcontroller/DeleteGameLogic.java
@@ -17,8 +17,11 @@ public class DeleteGameLogic {
 
     public String deleteGame() {
 
-        int gameId = screen.getMainMenuPanel().getDeleteGamePanel().getGameToDelete().getGameId();
+        if (screen.getMainMenuPanel().getDeleteGamePanel().getGameToDelete() != null) {
+            int gameId = screen.getMainMenuPanel().getDeleteGamePanel().getGameToDelete().getGameId();
 
-        return Api.deleteGame(gameId);
+            return Api.deleteGame(gameId);
+        }
+        return "";
     }
 }
